#4 Friday's Prototype Level
Made some quick and dirty doors and pressure plates, doors implement
OSControllable. Just slapped on the default character controller for the
dog. Haven't implemented the hiding spot, not sure yet how the proximity
paradox is implemented so I just laid down a slab of concrete so the
"dog" can get out of range