Tweaks
Added two booleans
In TimeManager script there is an Enable Rewind To Zero. Set that to
true to have the player automatically rewind the farthest.
Implementation is NOT robust... you have to keep the crouch button down.
If you get up mid rewind you might be able to force stop the thing...
Anyway I'll make it more robust if we make it the de facto way to
rewind.
Also added a boolean to the clone time attachment, set this to true if
you want the pup to create paradoxes with the clones. Look into the move
while crouching bug... can't seem to reproduce it and I really don't
know what could be causing it.