WarpIn Bubble
- Implemented warp in/out effect
The bubble could look better, but the goal was to create an object that
isn't a particle effect. Its state is affected by the time stop and the
rewind mechanics. Right now there isn't a bubble appearing when the
active player warps in but that shouldn't be too hard to do. The hard
part was figuring out how to make the engine deal with this time-synced
object.
- Camera snapping to clones
The camera will always try to snap to the most "recent" clone. Since any
clone can warp in at time !=0, the engine will display the most recent
clone currently active in the rewind state. If this doesn't make sense,
try doing the proto demo with the "pro" solution (the one that uses the
flash rewind) and you'll see what I mean