Mostly Done
I merged dev into PA_Tweaks, but I haven't merged "down" to dev yet.
Most conflicts should be resolved. Here's a recap of what's been done on
this branch.
- 3 rewind types selectable from TimeManager script
- Fixed clone color code
- Snapping camera to earliest clone
- TimeManager script requires a reference to the camera assigned to the
player
- Everything is hardcoded to display 0 (if needed could put a field
eventually that allows us to set this at will)
- Warp bubble
- not applied to main character yet, will do this ASAP but right now I
got other stuff to do
- there is a lifetime field in TimeManager and WarpBubble script, set
those two fields to the same value if it's not done already (should be
20 I think). Will hardcode something once we settle on a value
- The warp effect is a prefab whose state is modified by the engine's
clock, so it should look/behave as expected when doing stuff "out of
time". It's not super pretty but the implementation is robust and if we
want to improve we could swap it with something else, as long as it is
consistent with the actual interface
- Pressure plate
- All bugs related to pressure plates should be fixed
- Clones will toggle off pressure plate when warping out
- Plates will no longer trigger when disabling the LayerCollisions
- Added a bool check for robustness
As mentioned in #programmers, the lerp for the camera isn't done yet, it
will require some reworking of the engine to make it look nice