Ermergerd so much duct tape - poof text
Ugly as fuck, but took the time to fix anything that was looking weird
or unusual with the WarpIn bubble prefab. I'm not too proud of myself,
there is a lot, A LOT, of bad design choices in there.
Basically, the warpin effects appear/disappear when they are supposed
to. Regarding what they look like... that's another story